// src/components/pages/Report/ReportBmppIndividualDPOP/DetailPage/components/Detail/Detail.types.ts
export type DetailProps = {
  data?: {
    customerId?: string;
    cif?: string;
    customerName?: string;
    customerStatus?: string;
    customerType?: string;
    customerRelation?: string;
    divisionName?: string;
    modalEkuitas?: string;
    totalFasilitasExistingDebitur?: string;
    persyaratanBmppIndividual?: string;
    presentaseRealisasiBmppIndividual?: string;
    kesimpulan?: string;
    informasiPenarikanDataDataModalPer?: string;
    informasiPenarikanDataTanggalInputDataModal?: string;
    informasiPenarikanDataFasilitasExisting?: string;
    informasiPenarikanDataUsulanFasilitas?: string;
    dataBmppBy?: string;
    createdDate?: string;
  };
  isLoading: boolean;
};